Refactor Frame and FrameBatch into their own source file
